:sparkles: (Day 1 - Part 1) Solution using binary search tree and list
Using a simple binary search tree, I was able to traverse the list of values and quickly search for a corresponding value in the set to make up the target sum. Ideally, I could have traversed the tree rather than storing a separate structure exclusively for the benefit of easy traversal.